# specs/001-v1-marketplace/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Marathon Photo Marketplace API
  version: 1.0.0
  description: |
    API for the V1 Core Marketplace - a two-sided photo marketplace where runners
    search for race photos by bib number and purchase watermark-free downloads.

    ## Authentication
    - Public endpoints (search, checkout) require no authentication
    - Photographer endpoints require Supabase Auth JWT
    - Webhook endpoints use Stripe signature verification

    ## Rate Limiting
    - Search: 60 requests/minute per IP
    - Upload: 100 requests/minute per authenticated user
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: http://localhost:3000/api
    description: Development
...
paths:
...
  /search:
    get:
      operationId: searchPhotos
      summary: Search photos by bib number
      description: |
        Search for race photos by bib number. No authentication required (FR-001).
        Results are watermarked previews (FR-002).
      tags: [Search]
      parameters:
        - name: bib
          in: query
          required: true
          description: Bib number to search for
          schema:
            type: string
            minLength: 1
            maxLength: 10
            example: "12345"
        - name: event_id
          in: query
          description: Filter by specific event
          schema:
            type: string
            format: uuid
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
            enum: [quality, recent]
            default: quality
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: cursor
          in: query
          description: Pagination cursor
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /photos/{id}:
    get:
      operationId: getPhoto
      summary: Get photo details
      description: Get photo details including watermarked preview
      tags: [Photos]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Photo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhotoDetail'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== CHECKOUT ====================
  /checkout/session:
    post:
      operationId: createCheckoutSession
      summary: Create checkout session
      description: |
        Create a Stripe Checkout session for guest checkout (FR-003).
        No authentication required.
      tags: [Checkout]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutRequest'
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /downloads/{transactionId}:
    get:
      operationId: getDownloads
      summary: Get download links
      description: |
        Get download links for purchased photos (FR-004, FR-005).
        Validates purchase before providing signed URLs.
      tags: [Checkout]
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: email
          in: query
          required: true
          description: Email used for purchase (for guest verification)
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Download links
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadLinksResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== EVENTS ====================
...
  /webhooks/stripe:
    post:
      operationId: handleStripeWebhook
      summary: Handle Stripe webhook
      description: |
        Receive and process Stripe webhook events.
        Handles: checkout.session.completed, charge.refunded, charge.dispute.created,
        account.updated, payout.paid, payout.failed
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Invalid signature
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth JWT token
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Bib number is required"
        details:
          type: object
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
          nullable: true
        has_more:
          type: boolean

    # ==================== SEARCH ====================
    SearchResponse:
      type: object
      required: [photos, pagination]
      properties:
        bib_number:
          type: string
        event:
          $ref: '#/components/schemas/EventSummary'
        photos:
          type: array
          items:
            $ref: '#/components/schemas/PhotoPreview'
        pagination:
          $ref: '#/components/schemas/Pagination'
    PhotoPreview:
      type: object
      required: [id, thumbnail_url, watermarked_url, price_cents]
      properties:
        id:
          type: string
          format: uuid
        thumbnail_url:
          type: string
          format: uri
        watermarked_url:
          type: string
          format: uri
        event_id:
          type: string
          format: uuid
        event_name:
          type: string
        event_date:
          type: string
          format: date
        quality_score:
          type: number
          minimum: 0
          maximum: 1
        price_cents:
          type: integer
          example: 100
        bib_confidence:
          type: number
          minimum: 0
          maximum: 1
    PhotoDetail:
      allOf:
        - $ref: '#/components/schemas/PhotoPreview'
        - type: object
          properties:
            width:
              type: integer
            height:
              type: integer
            photographer_id:
              type: string
              format: uuid
            bibs:
              type: array
              items:
                $ref: '#/components/schemas/BibDetection'
    BibDetection:
      type: object
      properties:
        bib_number:
          type: string
        confidence:
          type: number
        is_verified:
          type: boolean

    # ==================== CHECKOUT ====================
    CreateCheckoutRequest:
      type: object
      required: [photo_ids, email]
      properties:
        photo_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
          maxItems: 50
        email:
          type: string
          format: email
        success_url:
          type: string
          format: uri
        cancel_url:
          type: string
          format: uri
    CheckoutSessionResponse:
      type: object
      required: [session_id, checkout_url]
      properties:
        session_id:
          type: string
        checkout_url:
          type: string
          format: uri
        amount_cents:
          type: integer
        photo_count:
          type: integer
        bundle_applied:
          type: string
          enum: [single, pack_5, pack_10, pack_20]
          nullable: true
...
    DownloadLinksResponse:
      type: object
      required: [transaction_id, downloads]
      properties:
        transaction_id:
          type: string
          format: uuid
        downloads:
          type: array
          items:
            type: object
            properties:
              photo_id:
                type: string
                format: uuid
              download_url:
                type: string
                format: uri
              expires_at:
                type: string
                format: date-time

    # ==================== EVENTS ====================
    EventSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        date:
          type: string
          format: date
        location:
          type: string
        photo_count:
          type: integer
...